Extract redis connection logic out of input class
